Move 'chat_history' variable out of the pipeline class.
create conversation.py
